# memcached performance tweaks, zstd usage