--- a/zerver/lib/display_recipient.py
+++ b/zerver/lib/display_recipient.py
@@ -34,6 +34,9 @@ def get_display_recipient_cache_key(
     return display_recipient_cache_key(recipient_id)
 
 
+# Note that the _same_ cache key is used for streams, which contain a
+# string, not a list[UserDisplayRecipient]!  This works because the
+# recipient space is distinct between the two.
 @cache_with_key(get_display_recipient_cache_key, timeout=3600 * 24 * 7)
 def get_display_recipient_remote_cache(
     recipient_id: int, recipient_type: int, recipient_type_id: int | None
@@ -121,6 +124,7 @@ def bulk_fetch_stream_names(
         cache_transformer=get_name,
         setter=lambda obj: obj,
         extractor=lambda obj: obj,
+        pickled_tupled=False,
     )
 
     return stream_display_recipients