--- a/zerver/lib/zstd_level9.py
+++ b/zerver/lib/zstd_level9.py
@@ -1,0 +1,10 @@
+from zstd import compress as original_compress
+from zstd import decompress
+
+__all__ = ["compress", "decompress"]
+
+
+def compress(data: bytes, level: int | None = None) -> bytes:
+    if level is None:
+        level = 9
+    return original_compress(data, level)