This slightly reduces the access times for get_members with a hot cache -- on a 6000-user realm, p50 decreases from 521ms to 499ms, with a similar drop in p90.  The memcache storage space drops from 108288 bytes to 67168, saving 38%.  Real-life savings are expected to be closer to 10% -- this is a test realm with users which are all somewhat similar.

<img width="600" height="371" alt="get_users access times" src="https://github.com/user-attachments/assets/67721024-0747-4edd-937b-08a000efc8a7" />

get_messages access times for hot caches are unchanged, but for cold caches, p50 for a 1400-message fetch drops from 581ms to 520ms.  The outliers are quite notable (especially as this is single-threaded and without nginx involved) especially as the time does not appear to be in memcached or postgres.  They're not new, or affected notably by this change, however.

<img width="600" height="371" alt="get_messages access times, with hot caches" src="https://github.com/user-attachments/assets/ecd58fe7-be67-4242-b284-9eb0e1c195f8" />
<img width="600" height="371" alt="get_messages access times, with cold caches" src="https://github.com/user-attachments/assets/b7060ca4-66aa-404e-bf5b-0b99a8cb1292" />